refactor(infra): migrate to postgres and reorganize user packages (#3)
