--- src/test/java/com/lernix/DemoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.lernix;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
